backend/requirements.txt
astroid==2.12.12
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dill==0.3.5.1
Flask==2.2.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
Frozen-Flask==0.18
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==5.0.0
install==1.3.5
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
mccabe==0.7.0
mysql-connector-python==8.0.31
mysqlclient==2.1.1
packaging==21.3
platformdirs==2.5.2
protobuf==3.20.1
pycodestyle==2.9.1
pylint==2.15.5
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.42
SQLAlchemy-Utils==0.38.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
twitter==1.18.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
